+++ System_Files/system.global.include.content.php
@@ -7,9 +7,9 @@ function content() {
         print('<script src="src/FOIL-time.js"></script>');
         print('<center><FOIL-font-size-40> Menu </FOIL-font-size-40></center>');
         print('<center>');
-        print('<a href="#" class="FOIL-button-regular-no-shadow"></a>');
-        print('<a href="#" class="FOIL-button-regular-no-shadow"></a>');
-        print('<a href="#" class="FOIL-button-regular-no-shadow"></a>');
+        print('<a href="#" class="FOIL-button-regular-no-shadow">Settings</a>');
+        print('<a href="./Views/Applications-Menu/index.php" class="FOIL-button-regular-no-shadow">Apps</a>');
+        print('<a href="./Views/Subjects/index.php" class="FOIL-button-regular-no-shadow">Subjects</a>');
         print("</center>");
 }
 content();