+++ System_Files/system.global.include.ui.php
@@ -1,2 +1,3 @@
 <link rel="stylesheet" href="../src/FOIL-UI.css" type="text/css">
 <link rel="stylesheet" href="../src/config.css" type="text/css">
+<link rel="stylesheet" href="../src/UI.css" type="text/css">