+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 
 - Added UI.css in [src/UI.css]. (The classic UI.css is back)
 
+- Include [src/UI.css] in [System_Files/system.global.include.ui.php]
+
 
 # v1.2.0
 - Changed layout padding from 40px (ALL) to 200px (ALL)