+++ CHANGELOG.md
@@ -1,11 +1,38 @@
 # CHANGELOG 
 
-# v.1.2.0
+# v1.3.0
+- Allow button padding to be set only in HTML syntax without wrapping in the src/config.css file
+
+**Example**:
+
+`<a href="#" class="button" style="padding 20px 20px 20px 20px">hello</a>`
+
+
+- Named buttons in the Home Screen. 
+
+- Created Views Folder
+
+- Created [src/position-text.css]
+
+- Created Subjects View in [Views/Subjects]
+
+- Created Applications View in [Views/Applications-Menu]
+
+- Updated Docs in [Views/README.txt]
+
+- Named button: [Math] in the Subjects View [Views/Subjects/index.php]
+
+- Added UI.css in [src/UI.css]. (The classic UI.css is back)
+
+- Include [src/UI.css] in [System_Files/system.global.include.ui.php]
+
+
+# v1.2.0
 - Changed layout padding from 40px (ALL) to 200px (ALL)
 - Removed For loop for adjusting layouts in Home Screen [System_Files/system.global.include.content.php]
 - Added Blank Buttons in the Home Screen. 
 
-# v.1.1.1
+# v1.1.1
 - Created CHANGELOG.md